requirements.txt
pocsuite
nose
requests
virtualenv
pymongo
pwntools
